refactor(source): give every fetcher subpackage a controller-shaped layout
Match the elegance of pkg/controllers, where each subpackage carries a
predictable spine + focused side files:

  pkg/controllers/{helmrelease,kustomization,source}/
    controller.go   — the Controller type + Configure + Start + reconcile
    dispatch.go     — emit children to the store
    fingerprint.go  — content-addressed dedup
    substitute.go   — kustomize-specific var sub

The fetcher subpackages followed no convention. oci/ stuffed 640 LOC of
mixed concerns into oci.go; git/git.go was 401 LOC; bucket/bucket.go
was 305 LOC after PR #407's file fold. The fold made each individual
file fat — concerns no longer mapped to files.

Reshape every fetcher subpackage to the controller pattern:

  pkg/source/oci/
    fetcher.go   (81)  — Fetcher type, Fetch entry, authIdentity
    auth.go      (116) — TLS, registry config, credential store
    fetch.go     (262) — workhorse, cache-hit gate, artifact composer
    marker.go    (98)  — .flate-digest / .flate-verified slot markers
    resolve.go   (143) — ref parsing, semver tag picking, revision shape
    cosign.go    (302) — pre-existing
    layer.go     (355) — pre-existing

  pkg/source/git/
    fetcher.go   (292) — Fetcher type, Fetch entry, fetch + fetchViaMirror,
                          canUseMirror, finalize, authIdentity
    checkout.go  (97)  — checkoutRef + updateSubmodules
    marker.go    (43)  — .flate-git-revision slot marker + worktree HEAD
    auth.go      (96)  — pre-existing
    tls.go       (50)  — pre-existing
    ssh.go       (41)  — pre-existing
    resolve.go   (127) — pre-existing
    materialize.go(31) — pre-existing

  pkg/source/bucket/
    fetcher.go   (127) — Fetcher type, Fetch entry, authIdentity
    auth.go      (67)  — resolveCredentials + resolveTransport
    endpoint.go  (39)  — normalizeEndpoint + schemeFor (re-extracted)
    walk.go      (99)  — list + parallel download + revision hash
    traversal.go (30)  — pre-existing

Each subpackage's `fetcher.go` carries a package-doc file map so a
new reader sees the structure at the top. No behavior changes — pure
file reorganization. Tests stay in place; they reference package
symbols which are now spread across files but otherwise unchanged.

Note: endpoint.go re-emerges as a 39-LOC file (it was folded into
bucket.go by PR #407). The fold optimized for "no thin files" — this
refactor optimizes for "one concern per file" which is the criterion
that makes pkg/controllers readable. Both can be right at different
granularities; here the concern split wins.
